[bug 854898] Allow youtube embeds in the forums.
Moved the video hook from wiki.parser to the parent sumo.parser and
added tests.
