+++ media/less/forums.less
@@ -2,6 +2,7 @@
 @import 'vendors.less';
 
 @import 'includes/editor-tools.less';
+@import 'includes/wiki-common.less';
 
 #forum-title {
   font-size: 46px;
